refactor: fix react-refresh lint warnings
