+++ src/components/sections/FAQSection/FAQSection.tsx
@@ -5,7 +5,7 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from '@/components/ui/accordion';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 
 interface FAQItem {
   question: { en: string; ar: string; eg: string };
