+++ src/components/sections/SolutionsSection/SolutionsSection.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 
 const SolutionsSection: React.FC = () => {
   const { t } = useLanguage();
