+++ src/components/sections/TechnologiesSection/TechnologiesSection.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 import {
   Carousel,
   CarouselContent,
