+++ src/components/ui/enhanced-button.tsx
@@ -1,51 +1,9 @@
 import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
-import { cva, type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 import { LoadingSpinner } from './loading-spinner';
-
-const buttonVariants = cva(
-  'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-all duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 relative overflow-hidden',
-  {
-    variants: {
-      variant: {
-        default:
-          'bg-primary text-primary-foreground hover:bg-primary-hover shadow-md hover:shadow-elegant transform hover:-translate-y-0.5 active:translate-y-0',
-        destructive:
-          'bg-destructive text-destructive-foreground hover:bg-destructive-hover shadow-md hover:shadow-elegant',
-        outline:
-          'border border-input bg-background hover:bg-accent hover:text-accent-foreground shadow-sm hover:shadow-md',
-        secondary:
-          'bg-secondary text-secondary-foreground hover:bg-secondary-hover shadow-sm hover:shadow-md',
-        ghost: 'hover:bg-accent hover:text-accent-foreground',
-        link: 'text-primary underline-offset-4 hover:underline',
-        premium:
-          'bg-gradient-to-r from-primary to-primary-hover text-primary-foreground shadow-premium hover:shadow-glow transform hover:-translate-y-1 active:translate-y-0',
-        glow: 'bg-primary text-primary-foreground shadow-glow hover:shadow-glow-lg animate-glow',
-      },
-      size: {
-        default: 'h-10 px-4 py-2',
-        sm: 'h-9 rounded-md px-3',
-        lg: 'h-11 rounded-md px-8 text-base',
-        xl: 'h-14 rounded-lg px-10 text-lg',
-        icon: 'h-10 w-10',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-      size: 'default',
-    },
-  },
-);
-
-export interface EnhancedButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  loading?: boolean;
-  leftIcon?: React.ReactNode;
-  rightIcon?: React.ReactNode;
-}
+import type { EnhancedButtonProps } from './enhanced-button-utils';
+import { buttonVariants } from './enhanced-button-utils';
 
 const EnhancedButton = React.forwardRef<HTMLButtonElement, EnhancedButtonProps>(
   (
@@ -83,4 +41,4 @@ const EnhancedButton = React.forwardRef<HTMLButtonElement, EnhancedButtonProps>(
 
 EnhancedButton.displayName = 'EnhancedButton';
 
-export { EnhancedButton, buttonVariants };
+export { EnhancedButton };
