+++ src/components/ui/sonner-utils.ts
@@ -1,0 +1,1 @@
+export { toast } from 'sonner';
