+++ src/components/ui/sonner.tsx
@@ -1,5 +1,6 @@
 import { useTheme } from 'next-themes';
-import { Toaster as Sonner, toast } from 'sonner';
+import { Toaster as Sonner } from 'sonner';
+import { toast } from './sonner-utils';
 
 type ToasterProps = React.ComponentProps<typeof Sonner>;
 
@@ -26,4 +27,4 @@ const Toaster = ({ ...props }: ToasterProps) => {
   );
 };
 
-export { Toaster, toast };
+export { Toaster };
