+++ src/contexts/LanguageContext.helpers.ts
@@ -1,0 +1,23 @@
+import { createContext, useContext } from 'react';
+import type { Language, TranslationKey } from '@/locales';
+
+export interface LanguageContextType {
+  language: Language;
+  setLanguage: (lang: Language) => void;
+  t: (key: TranslationKey) => string;
+  isRTL: boolean;
+}
+
+export const LanguageContext = createContext<LanguageContextType | undefined>(
+  undefined,
+);
+
+export const useLanguage = () => {
+  const context = useContext(LanguageContext);
+  if (!context) {
+    throw new Error('useLanguage must be used within a LanguageProvider');
+  }
+  return context;
+};
+
+export type { Language, TranslationKey };
