+++ src/contexts/LanguageContext.tsx
@@ -1,24 +1,12 @@
-import {
-  createContext,
-  useContext,
-  useState,
-  useEffect,
-  ReactNode,
-} from 'react';
+import { useState, useEffect, ReactNode } from 'react';
 import i18n from '@/lib/i18n';
 import { useTranslation } from 'react-i18next';
 import { type Language, type TranslationKey } from '@/locales';
-
-interface LanguageContextType {
-  language: Language;
-  setLanguage: (lang: Language) => void;
-  t: (key: TranslationKey) => string;
-  isRTL: boolean;
-}
-
-const LanguageContext = createContext<LanguageContextType | undefined>(
-  undefined,
-);
+import {
+  LanguageContext,
+  useLanguage,
+  type LanguageContextType,
+} from './LanguageContext.helpers';
 
 export const LanguageProvider = ({ children }: { children: ReactNode }) => {
   // Default language Arabic with RTL direction
@@ -52,13 +40,3 @@ export const LanguageProvider = ({ children }: { children: ReactNode }) => {
     </LanguageContext.Provider>
   );
 };
-
-export const useLanguage = () => {
-  const context = useContext(LanguageContext);
-  if (!context) {
-    throw new Error('useLanguage must be used within a LanguageProvider');
-  }
-  return context;
-};
-
-export type { Language, TranslationKey };
