+++ src/pages/CareerPostPage.tsx
@@ -5,7 +5,7 @@ import { useQuery } from '@tanstack/react-query';
 import Navigation from '@/components/layout/Navigation';
 import Footer from '@/components/layout/Footer';
 import BackToTopButton from '@/components/layout/BackToTopButton';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 import { fetchCareerPost, type CareerPost } from '@/lib/cms';
 
 const CareerPostPage: React.FC = () => {
