+++ src/pages/PartnerPage.tsx
@@ -5,7 +5,7 @@ import { useQuery } from '@tanstack/react-query';
 import Navigation from '@/components/layout/Navigation';
 import Footer from '@/components/layout/Footer';
 import BackToTopButton from '@/components/layout/BackToTopButton';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 import { fetchPartner, type Partner } from '@/lib/cms';
 
 const PartnerPage: React.FC = () => {
