+++ src/pages/SiteIndex.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Helmet } from 'react-helmet';
 import Index from './Index';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 
 const SiteIndex: React.FC = () => {
   const { t } = useLanguage();
