+++ src/pages/SolutionsPage.tsx
@@ -6,7 +6,7 @@ import CTASection from '@/components/sections/CTASection/CTASection';
 import ContactSection from '@/components/sections/ContactSection/ContactSection';
 import Footer from '@/components/layout/Footer';
 import BackToTopButton from '@/components/layout/BackToTopButton';
-import { useLanguage } from '@/contexts/LanguageContext';
+import { useLanguage } from '@/contexts/LanguageContext.helpers';
 
 const SolutionsPage: React.FC = () => {
   const { t } = useLanguage();
